+++ common/cards/attach/armor.ts
@@ -1,113 +1,168 @@
 import {CardComponent, ObserverComponent} from '../../components'
-import {ExpansionT} from '../../const/expansions'
 import {GameModel} from '../../models/game-model'
-import {CardRarityT, TokenCostT} from '../../types/cards'
 import {beforeAttack} from '../../types/priorities'
 import {attach} from '../defaults'
 import {Attach} from '../types'
 
-function getArmor(
-	props: {
-		id: string
-		name: string
-		rarity: CardRarityT
-		numericId: number
-		tokens: TokenCostT
-		expansion: ExpansionT
-	},
+function armorBlockDamage(
 	amount: number,
-): Attach {
-	return {
-		...attach,
-		id: props.id,
-		numericId: props.numericId,
-		name: props.name,
-		expansion: props.expansion,
-		rarity: props.rarity,
-		tokens: props.tokens,
-		description: `When the Hermit this card is attached to takes damage, that damage is reduced by up to ${amount}hp each turn.`,
-		onAttach(
-			game: GameModel,
-			component: CardComponent,
-			observer: ObserverComponent,
-		) {
-			const {player, opponentPlayer} = component
+	game: GameModel,
+	component: CardComponent,
+	observer: ObserverComponent,
+) {
+	const {player, opponentPlayer} = component
 
-			let damageBlocked = 0
+	let damageBlocked = 0
 
-			observer.subscribeWithPriority(
-				game.hooks.beforeAttack,
-				beforeAttack.EFFECT_REDUCE_DAMAGE,
-				(attack) => {
-					if (!attack.isTargeting(component) || attack.isType('status-effect'))
-						return
+	observer.subscribeWithPriority(
+		game.hooks.beforeAttack,
+		beforeAttack.EFFECT_REDUCE_DAMAGE,
+		(attack) => {
+			if (!attack.isTargeting(component) || attack.isType('status-effect'))
+				return
 
-					if (damageBlocked < amount) {
-						const damageReduction = Math.min(
-							attack.calculateDamage(),
-							amount - damageBlocked,
-						)
-						damageBlocked += damageReduction
-						attack.addDamageReduction(component.entity, damageReduction)
-					}
-				},
-			)
-
-			const resetCounter = () => {
-				damageBlocked = 0
+			if (damageBlocked < amount) {
+				const damageReduction = Math.min(
+					attack.calculateDamage(),
+					amount - damageBlocked,
+				)
+				damageBlocked += damageReduction
+				attack.addDamageReduction(component.entity, damageReduction)
 			}
-
-			// Reset counter at the start of every turn
-			observer.subscribe(player.hooks.onTurnStart, resetCounter)
-			observer.subscribe(opponentPlayer.hooks.onTurnStart, resetCounter)
 		},
+	)
+
+	const resetCounter = () => {
+		damageBlocked = 0
 	}
+
+	// Reset counter at the start of every turn
+	observer.subscribe(player.hooks.onTurnStart, resetCounter)
+	observer.subscribe(opponentPlayer.hooks.onTurnStart, resetCounter)
 }
 
-export const GoldArmor = getArmor(
-	{
-		id: 'gold_armor',
-		numericId: 29,
-		name: 'Gold Armour',
-		expansion: 'default',
-		rarity: 'common',
-		tokens: 0,
-	},
-	10,
-)
+function armorBlockEffect(
+	game: GameModel,
+	component: CardComponent,
+	observer: ObserverComponent,
+) {
+	observer.subscribeWithPriority(
+		game.hooks.beforeAttack,
+		beforeAttack.EFFECT_BLOCK_DAMAGE,
+		(attack) => {
+			if (!attack.isTargeting(component)) {
+				return
+			}
 
-export const IronArmor = getArmor(
-	{
-		id: 'iron_armor',
-		numericId: 45,
-		name: 'Iron Armour',
-		expansion: 'default',
-		rarity: 'common',
-		tokens: 2,
-	},
-	20,
-)
+			// only protect against su attacks and attacks which have been redirected by su cards
+			let suRedirect = false
 
-export const DiamondArmor = getArmor(
-	{
-		id: 'diamond_armor',
-		numericId: 13,
-		name: 'Diamond Armour',
-		expansion: 'default',
-		rarity: 'rare',
-		tokens: 3,
-	},
-	30,
-)
+			const lastTargetChange = attack.getHistory('redirect').pop()
+			if (lastTargetChange) {
+				// This attack has been redirected to us by a su card
+				suRedirect = true
+			}
 
-export const NetheriteArmor = getArmor(
-	{
-		id: 'netherite_armor',
-		numericId: 82,
-		name: 'Netherite Armour',
-		expansion: 'default',
-		rarity: 'ultra_rare',
-		tokens: 4,
+			if (attack.isType('effect') || suRedirect) {
+				attack.multiplyDamage(component.entity, 0).lockDamage(component.entity)
+			}
+		},
+	)
+}
+
+export const GoldArmor = {
+	...attach,
+	id: 'gold_armor',
+	numericId: 29,
+	name: 'Gold Armour',
+	expansion: 'default',
+	rarity: 'common',
+	tokens: 0,
+	description:
+		'When the Hermit this card is attached to takes damage, that damage is reduced by up to 10hp each turn.',
+	onAttach(
+		game: GameModel,
+		component: CardComponent,
+		observer: ObserverComponent,
+	) {
+		return armorBlockDamage(10, game, component, observer)
 	},
-	40,
-)
+}
+
+export const IronArmor = {
+	...attach,
+	id: 'iron_armor',
+	numericId: 45,
+	name: 'Iron Armour',
+	expansion: 'default',
+	rarity: 'common',
+	tokens: 2,
+	description:
+		'When the Hermit this card is attached to takes damage, that damage is reduced by up to 20hp each turn.',
+	onAttach(
+		game: GameModel,
+		component: CardComponent,
+		observer: ObserverComponent,
+	) {
+		return armorBlockDamage(20, game, component, observer)
+	},
+}
+
+export const ChainmailArmor: Attach = {
+	...attach,
+	id: 'chainmail_armor',
+	numericId: 119,
+	name: 'Chainmail Armour',
+	expansion: 'alter_egos',
+	rarity: 'common',
+	tokens: 1,
+	description:
+		'Prevents any damage from effect cards and any damage redirected by effect cards to the Hermit this card is attached to.',
+	onAttach(
+		game: GameModel,
+		component: CardComponent,
+		observer: ObserverComponent,
+	) {
+		armorBlockEffect(game, component, observer)
+	},
+}
+
+export const DiamondArmor: Attach = {
+	...attach,
+	id: 'diamond_armor',
+	numericId: 13,
+	name: 'Diamond Armour',
+	expansion: 'default',
+	rarity: 'rare',
+	tokens: 3,
+	description:
+		'Prevent up to 20HP damage taken. Also prevents damage to this Hermit from all effect cards.',
+	onAttach(
+		game: GameModel,
+		component: CardComponent,
+		observer: ObserverComponent,
+	) {
+		armorBlockEffect(game, component, observer)
+		armorBlockDamage(20, game, component, observer)
+	},
+}
+
+export const NetheriteArmor: Attach = {
+	...attach,
+	id: 'netherite_armor',
+	numericId: 82,
+	name: 'Netherite Armour',
+	expansion: 'default',
+	rarity: 'ultra_rare',
+	tokens: 4,
+	description:
+		'Prevent up to 20HP damage taken. Also prevents damage to this Hermit from all effect cards. Opponent can not make this Hermit go AFK.',
+	onAttach(
+		game: GameModel,
+		component: CardComponent,
+		observer: ObserverComponent,
+	) {
+		armorBlockEffect(game, component, observer)
+		armorBlockDamage(20, game, component, observer)
+	},
+}
